CassandraQuery Class
Got pretty far on the CassandraQuery class. Still some errors in the handling of keyspaces (i.e. it doesn't create them yet). Keyspace creation is the next step and should be automagically handled by the DatabaseWrapper (connection) class.

Added a where.py which subclasses the Django WhereNode and EverythingNode. I'm hoping that due to the stark similarities between SQL and CQL we can mostly reuse these classes to construct our where expression trees.

I also tested the cluster configuration and connection and was able to get it to connect to a Cassandra cluster running on my local machine.

Not ready for a merge into master yet.