Keyspace and Database Creation
Doubled back on the DatabaseWrapper class and made this guy for creating keyspaces. This happens automatically if the keyspace does not already exist. Should probably figure out some way to alter a keyspace but I can't remember if Cassandra even allows this without dropping the whole thing.

Implemented table creation in the DatabaseCreation class. For some reason the method for creating a database isn't getting hit by Django during syncdb. Need to figure out what's going on there still. Also still need to define the type of the column based on the type of the field. This consists of defining the mapping of field type names to Cassandra column data types.