Select/Delete Filters
Got the built in Django WhereNode class generating our where statements.

Now there's an issue with some of the built in Django applications (django contenttypes to be specific) trying to filter on non-indexed columns. The workaround here might have to be pulling in the whole table and filtering in memory which sucks but I belive that's the solution our current database backend is using. This shouldn't be an issue with small tables like contenttypes where there is only one entry per model and the dataset doesn't grow but a warning should certainly be raised in this event. I'll investigate alternate workarounds to this issue.