Support For Inefficient Filtering of Results
Some utilities for providing support for inefficient filtering. There are some parts of django and third party apps that require this to function properly. There should probably be a warning thrown in the case of using an filter that is required to be processed in memory and only should ever be used on lightweight datasets.