CQLEngine Library Integration
Initially it seemed silly to use a two ORM solution but it turns out I can almost completely obfuscate the cqlengine inside of the database backend. This turns out to be rather simple with a little meta-programming since the Django ORM and the cqlengine ORM are so similar.

So now the database backend leverages cqlengine for actual database management and queries and the results from this are returned in a form that django expects. So far this is working with stock Django 1.7 in my tests. Haven't tested the full ORM tests suite on it yet but hopefully there won't be too many headaches.