Cassandra Cursor
I made a CassandraCursor class that implements a lot of the same interface as the MySQLdb cursor class so that django can pass this object around and that our database backend can find this useful. This will make a lot of the database introspection and migration features that are apart of Django 1.7

I also noticed that Django 1.7 refactored a lot of the DatabaseWrapper functionality into methods that make sense. I already had this logically seperated in a similar way but I went ahead and refactored what I had into the new DatabaseWrapper interface methods.