Range Filtering, Deleting and Counting Improvements
There was an error in how the backend determines wether or not a range query can be evaluated efficiently by Cassandra. It was allowing attempts of efficent filtering on non-clustering columns.

Had some errors in how I was handling deleting and counting methods. I thought the queries were build at that point in code execution. I was mistaken but now the queries are build correctly and the results are as expected.