Inefficent Queries (get/filter) Don't Return Results In Some Cases
Nasty bug here where for inefficient queries the limit on the query wasn't being reset. Somewhere in the get/filter django sets a limit to 21 (I think this is MAX_GET_RESULTS +1; use grep to find if curious). When there were more than 21 results in the query set there was a good chance that the result that matches the provided filter would not be in the result set.