Partition/Cluster Primary Key Models With Mutltiple Fields Fail on Insert
There was an issue where Django decides to do an update instead of an insert which causes a malformed CQL query on the backend. I made a change to ColumnFamily model save method where it always passes force_insert=True. ColumnFamily is now required to utilize partition and clustering keys.

Added some tests to cover this.